[IMP]changes to avoid outside top level imports and mock student module
