+++ eventtracking/backends/routing.py
@@ -2,15 +2,12 @@
 
 from __future__ import absolute_import
 
-import json
 import logging
 from collections import OrderedDict
 
 import six
 
 from eventtracking.processors.exceptions import EventEmissionExit
-from eventtracking.helpers import GenericJSONEncoder
-from eventtracking.tasks import async_send
 
 
 LOG = logging.getLogger(__name__)
@@ -138,37 +135,3 @@ class RoutingBackend:
                 LOG.exception(
                     'Unable to send event to backend: %s', name
                 )
-
-
-class AsyncRoutingBackend(RoutingBackend):
-    """
-    Route events to configured backends asynchronously
-    """
-
-    def send(self, event):
-        """
-        Process the event using all registered processors and send it to all registered backends.
-        """
-        try:
-            processed_event = self.process_event(event)
-        except EventEmissionExit:
-            return
-
-        try:
-            json_event = json.dumps(processed_event, cls=GenericJSONEncoder)
-        except ValueError:
-            LOG.exception(
-                'JSONEncodeError: Unable to encode event:%s', processed_event
-            )
-            return
-
-        for name, backend in six.iteritems(self.backends):
-            try:
-                json_backend = json.dumps(backend, cls=GenericJSONEncoder)
-            except ValueError:
-                LOG.exception(
-                    'JSONEncodeError: Unable to encode backend: %s', name
-                )
-                continue
-
-            async_send.delay(json_backend, json_event, backend_name=name)
