+++ eventtracking/backends/tests/test_routing.py
@@ -9,7 +9,8 @@ import six
 from mock import MagicMock, call, sentinel, patch
 from six.moves import range
 
-from eventtracking.backends.routing import RoutingBackend, AsyncRoutingBackend
+from eventtracking.backends.routing import RoutingBackend
+from eventtracking.backends.async_routing import AsyncRoutingBackend
 from eventtracking.processors.exceptions import EventEmissionExit
 
 
@@ -303,7 +304,7 @@ class TestAsyncRoutingBackend(TestCase):
             str(i): MagicMock() for i in range(3)
         }
 
-    @patch('eventtracking.backends.routing.async_send')
+    @patch('eventtracking.backends.async_routing.async_send')
     def test_event_emission_exit(self, mocked_async_send):
         mocked_processor = MagicMock(side_effect=EventEmissionExit)
         backend = AsyncRoutingBackend(processors=[
@@ -312,9 +313,9 @@ class TestAsyncRoutingBackend(TestCase):
         backend.send(self.sample_event)
         mocked_async_send.assert_not_called()
 
-    @patch('eventtracking.backends.routing.LOG')
-    @patch('eventtracking.backends.routing.json.dumps', side_effect=ValueError)
-    @patch('eventtracking.backends.routing.async_send')
+    @patch('eventtracking.backends.async_routing.LOG')
+    @patch('eventtracking.backends.async_routing.json.dumps', side_effect=ValueError)
+    @patch('eventtracking.backends.async_routing.async_send')
     def test_with_value_error_in_event_json_encoding(self, mocked_async_send, _, mocked_log):
 
         backend = AsyncRoutingBackend()
@@ -323,8 +324,8 @@ class TestAsyncRoutingBackend(TestCase):
         mocked_log.exception.assert_called_once_with('JSONEncodeError: Unable to encode event:%s', self.sample_event)
         mocked_async_send.assert_not_called()
 
-    @patch('eventtracking.backends.routing.LOG')
-    @patch('eventtracking.backends.routing.async_send.delay')
+    @patch('eventtracking.backends.async_routing.LOG')
+    @patch('eventtracking.backends.async_routing.async_send.delay')
     def test_with_value_error_in_one_backend_json_encoding(self, mocked_async_send, mocked_log):
         """Test the backend's send method when one of the three backends
         could not be encoded into json.
@@ -332,7 +333,7 @@ class TestAsyncRoutingBackend(TestCase):
 
         json_event = json.dumps(self.sample_event)
 
-        with patch('eventtracking.backends.routing.json.dumps') as mocked_json_dumps:
+        with patch('eventtracking.backends.async_routing.json.dumps') as mocked_json_dumps:
             mocked_json_dumps.side_effect = [json_event, ValueError, '{"mocked_backend":1}', '{"mocked_backend":2}']
             backend = AsyncRoutingBackend(backends=self.mocked_backends)
             backend.send(self.sample_event)
